requirements.txt
accelerate==0.24.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.1.1
controlnet-aux==0.0.7
cycler==0.12.1
Cython==0.29.36
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.14.0
docopt==0.6.2
einops==0.7.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.45.1
fsspec==2023.10.0
ftfy==6.1.3
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.1
numpy==1.24.4
nvdiffrast==0.3.1
opencv-python==4.8.1.78
packaging==23.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pip==23.3.1
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
prompt_toolkit==3.0.45
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyembree==0.2.11
Pygments==2.18.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2023.10.3
requests==2.31.0
rpds-py==0.18.1
Rtree==1.1.0
safetensors==0.4.1
scikit-image==0.21.0
scipy==1.10.1
setuptools==60.10.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tifffile==2023.7.10
timm==0.9.12
tinycss2==1.3.0
tokenizers==0.15.0
torch==2.1.0
torchtyping==0.1.4
torchvision==0.16.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
transformers==4.35.2
trimesh==3.23.5
triton==2.1.0
typeguard==4.1.5
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.12
webencodings==0.5.1
wheel==0.37.1
xatlas==0.0.8
xformers==0.0.22.post7
yarg==0.1.9
zipp==3.17.0
